lil bit of refactoring
fix comments and imports

rm commited file
