+++ example.vaults.config.json
