--- public/example.vaults.config.json
@@ -1,21 +1,0 @@
-[
-  {
-    "chainId": 31337,
-    "vaultId": "usdc",
-    "vaultAddress": "0x0DCd1Bf9A1b36cE34237eEaFef220932846BCD82",
-    "vaultProxyAddress": "0x9A676e781A523b5d0C0e43731313A708CB607508",
-    "tokenAddress": "0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9",
-    "transactionConfirmationTimeout": 60000,
-    "startBlock": 0,
-    "name": "USDC Stable Vault",
-    "description": "Stable yield generation for USDC deposits.",
-    "token": "USDC",
-    "aprRequest": {
-      "type": "api",
-      "url": "/api/mock-apr",
-      "method": "GET",
-      "headers": {},
-      "body": {}
-    }
-  }
-]
