+++ src/app/api/mock-apr/route.ts
@@ -5,7 +5,7 @@ export async function GET() {
   // In a real application, this would be fetched from a real data source
   // such as a database, external API, or blockchain contract
 
-  const mockApr = "0.0412";
+  const mockApr = "0.1920";
 
   // Add artificial delay to simulate network latency (optional)
   await new Promise((resolve) => setTimeout(resolve, 500));
