+++ src/app/vault/[id]/page.tsx
@@ -5,10 +5,10 @@ import { useViewAllVaults, useVaultContract, useTokenBalances } from "@/hooks";
 import { useAccount } from "wagmi";
 import { useState } from "react";
 import { isValidNumberInput } from "@/lib";
-import { Card, useToast } from "@/components";
+import { useToast } from "@/context";
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { QUERY_KEYS } from "@/const";
-import { Button, Input } from "@/components";
+import { Button, Input, Card } from "@/components";
 import { formatUnits } from "viem";
 
 export default function VaultPage({ params }: { params: { id: string } }) {
