+++ src/components/index.ts
@@ -1,11 +1,7 @@
-export * from "./AppKitProvider";
 export * from "./ConnectButton";
 export * from "./Button";
 export * from "./Card";
 export * from "./Footer";
 export * from "./Header";
-export * from "./ChainProviders";
-export * from "./Toast";
 export * from "./Input";
-export * from "./VaultsConfigProvider";
 export * from "./MobileDetection";
