+++ src/hooks/useViewAllVaults.ts
@@ -5,10 +5,9 @@ import { useAccount, useConfig } from "wagmi";
 import { QUERY_KEYS, valenceVaultABI } from "@/const";
 import { readContract, readContracts } from "@wagmi/core";
 import { erc20Abi } from "viem";
-import { VaultConfig } from "@/lib/readVaultsConfig";
 import { formatBigInt, fetchAprFromApi, fetchAprFromContract } from "@/lib";
 import { useQueries } from "@tanstack/react-query";
-import { useVaultsConfig } from "@/components";
+import { useVaultsConfig, type VaultConfig } from "@/context";
 
 export type VaultData = VaultConfig & {
   tokenDecimals: number;
