+++ src/lib/index.ts
@@ -2,3 +2,4 @@ export * from "./utils";
 export * from "./helper";
 export * from "./fetchVaultApr";
 export * from "./readVaultsConfig";
+export * from "./schema";
